some refactoring

+++ learner.h
@@ -36,8 +36,8 @@ public:
     void init(size_t replica_count, size_t _id);
     Message* update_vote(int n_a, int seq_num, std::string value);
     void print_log();
-    size_t get_seqnum();
-    size_t get_seqnum_with_skip();
+    int get_seqnum();
+    int get_seqnum_with_skip();
 };
 
 #endif /* learner_h */
